+++ src/components-temp/ThoughtForm.jsx
