+++ src/components-temp/ThoughtHeart.jsx
