+++ src/styles-temp/ThoughtForm.css
