+++ src/styles-temp/ThoughtHeart.css
